Move filters outside util
